Finally fixed the bug taht's causing pass through.
The problem is in CollisionDetector, circle to aabb detection. Where
when clamp result is a corner of the aabb, we already give the result a
non null value. However, we still need to determine the distance between
the player and the corner to verify whether there is collision or not.
This bug is an obvious logical bug, but took me a long time to find out.