[BREAK] This commit breaks build.
I want fuel indicator to be a more scarce resource, the player should
always be carefule when to use the fuel and where to steer their ship.
Thus, making the player think about what's the best strategy to finish a
level, the fuel left can be seen as a way to measure how good the player
did too.