[PLAYABLE] This version is playable. With an improved fuel bar.
The fuel bar now displays how many barrels of fuel you have, and for
current barrel how much fuel's left. The usage of the fuel is much
faster than before. There's an time that the player won't be able to
move the ship when the fuel is being replaced. Next I will need to make
that time visiable through the fuel bar indicator.