[WIP] Step 1 of level builder
First, we need to make each game object serializable to json, so that we
can use this json as the representation of our level, store them, and
make them loadable.

During the process, I decided to refactor a bit and finish what I
originally designed to separate collision_body with display_body.
This help makes the serialization process more explicit and clear.