Major Refactor in progress, may not be playable.
Like mentioned in last commit, physics and display should be isolated from
each other. When thinking about that, I found a major refactoring is
necessary. I'll need to separate the concept of physical body and it's
appearance. Appearance should be attached to a physical body, and render
based on the properties of the physical body. For example, rotation is
applied to physical body, but should be applied to appearance. The
appearance should look at the physical body and render accordingly.

Also added the work in progress `sample.json`, a draft of the json
representation of each level.