Refactor the rendering system for geometry shape rendering
Now it becomes a 2 layer design:
1. geometry/* this is where all the abstract geometry shapes are, they
can be used for collision detection or rendering since they are pure
abstract concepts
2. view/* this is where all the view classes are. They correspond to
each of the geometry shape in geometry/*. This is the layer that
encapuslates the information of style, each shape's color and fill will
be defined here. Also bound box is defined here.